Fix issue with relative import
